+++ packages/ubcli/lib/createStore.js
@@ -63,6 +63,7 @@ module.exports = function createStore (options) {
   }
 
   function createOneStore (cStore) {
+    let newStores=0
     if (!cStore.storeType) {
       cStore.storeType = 'FileSystem'
     }
@@ -74,9 +75,10 @@ module.exports = function createStore (options) {
       if (!fs.existsSync(cStorePath)) {
         console.log(`\tStore ${cStore.name}: path '${cStorePath}' not exists and will be created`)
         fs.mkdirSync(cStorePath)
+        newStores++
       }
     } else {
-      console.log('\tskipped - path not defined')
+      console.log('\tStore ${cStore.name}: skipped - path not defined')
     }
     if (cStore.tempPath) {
       const tmp = cStore.tempPath // already converted to absolute by argv
@@ -85,10 +87,15 @@ module.exports = function createStore (options) {
         fs.mkdirSync(tmp)
       }
     }
-    console.log('BLOB stores path is OK')
+    return newStores
   }
 
-  selectedStores.forEach(createOneStore)
+  const created = selectedStores.forEach(createOneStore)
+  if (created) {
+    console.log(`BLOB stores folders roots are OK (${created} new root folders are created)`)
+  } else {
+    console.log('BLOB stores folders roots are OK')
+  }
 }
 
 module.exports.shortDoc = `Create internal BLOB store structure (folders) for