+++ packages/ubcli/lib/linkStatic.js
@@ -223,15 +223,12 @@ goto :eof
   console.log(`
 
 ${WINDOWS ? 'CMD' : 'Bash'} script ${resFn} is created
-
-Review a script, take care about target folder and package list.
-
-In case some package should not be exposed to client add a section
-"config": {"ubmodel": {} } into corresponding package.json.
-
-Use a command:
-${WINDOWS ? '.\\.linkStatic.cmd' : 'chmod +x ./.linkStatic.sh && ./.linkStatic.sh'}
-to link a static`)
+  Review a script, take care about target folder and package list.
+  In case some package should not be exposed to client add a section
+        "config": {"ubmodel": {} } into corresponding package.json.
+  Use a command:
+    ${WINDOWS ? '.\\.linkStatic.cmd' : 'chmod +x ./.linkStatic.sh && ./.linkStatic.sh'}
+  to link a static`)
 
   // let pjsPath = path.join(cfgPath, 'package.json')
   // if (!fs.existsSync(pjsPath)) {